:bug: [ Fix ] - Fix Bootstrap inner DTO type

+++ src/main/java/com/gdsc/toplearth_server/application/dto/user/UpdateGoalResponseDto.java
@@ -1,10 +1,9 @@
 package com.gdsc.toplearth_server.application.dto.user;
 
-import java.math.BigDecimal;
 import lombok.Builder;
 
 @Builder
 public record UpdateGoalResponseDto(
-        BigDecimal goalDistance
+        Double goalDistance
 ) {
 }
